refactor(consumption): split gear_inference into models + clustering helpers
`lib/features/consumption/domain/services/gear_inference.dart` (481
LOC) exceeded the 300-LOC budget. Split via `part` into four
co-libraried files — same library scope, no public API surface
changes, behaviour bit-exact:

- `gear_inference.dart`               — 300 LOC (was 481): library
  docstring + public `inferGears`.
- `gear_inference_models.dart`        —  48 LOC: `InferredGearSample`,
  `GearInferenceResult` value classes.
- `_gear_clustering.dart`             —  69 LOC: k-means tuning
  constants + `_percentile` / `_nearestCentroid` helpers.
- `_gear_coaching_metrics.dart`       —  99 LOC: the
  `computeSecondsBelowOptimalGear` coaching-metric helper.

Pattern matches the existing `maintenance_analyzer.dart` ↔
`_maf_deviation_detector.dart` split in the same directory: leading-
underscore filename signals "library-internal" + `part of` keeps
private identifiers in scope without re-exports.

Public symbols (`inferGears`, `computeSecondsBelowOptimalGear`,
`InferredGearSample`, `GearInferenceResult`) all still resolve via
`import '.../gear_inference.dart'` — every existing import site
(trip_recording_controller.dart, the test files, etc.) compiles
unchanged. The 18 `gear_inference_test.dart` cases pass without
modification, as do the 11 dependent tests in
trip_recording_controller_gear_inference_test.dart and
driving_insights_card_low_gear_test.dart.

Refs #563 phase: gear_inference
